Integrate with our org's OpenAM IDP
This change addresses the issue of OpenAM's RP auth flow not allowing
multiple auth methods for the agent in one request and that the scope
field should not be in the access token call after the authorize step.
It also addresses that we don't have email in our scope and that the
audience is returned as an array.